Restructure package for simplicity
