+++ .isort.cfg
@@ -1,3 +1,3 @@
 [settings]
 profile = black
-known_first_party = fs.synapsefs
+known_first_party = synapsefs
