+++ Dockerfile
@@ -12,4 +12,4 @@ COPY src ./src/
 RUN SETUPTOOLS_SCM_PRETEND_VERSION_FOR_SAGETASKS=${PKG_VERSION} \
     python -m pip install .
 
-CMD [ "python", "-c", "import fs.synapsefs" ]
+CMD [ "python", "-c", "import synapsefs" ]
