+++ docs/conf.py
@@ -34,7 +34,7 @@ except ImportError:
     from sphinx import apidoc
 
 output_dir = os.path.join(__location__, "api")
-module_dir = os.path.join(__location__, "../src/fs.synapsefs")
+module_dir = os.path.join(__location__, "../src/synapsefs")
 try:
     shutil.rmtree(output_dir)
 except FileNotFoundError:
@@ -124,7 +124,7 @@ copyright = "2022, Bruno Grande"
 # If you don’t need the separation provided between version and release,
 # just set them both to the same value.
 try:
-    from fs.synapsefs import __version__ as version
+    from synapsefs import __version__ as version
 except ImportError:
     version = ""
