+++ setup.cfg
@@ -35,7 +35,7 @@ classifiers =
 
 [options]
 zip_safe = False
-packages = fs.synapsefs, fs.opener
+packages = find_namespace:
 include_package_data = True
 package_dir =
     =src
@@ -88,13 +88,13 @@ dev =
 [options.entry_points]
 # Add here console scripts like:
 # console_scripts =
-#     script_name = fs.synapsefs.module:function
+#     script_name = synapsefs.module:function
 # For example:
 # console_scripts =
-#     fibonacci = fs.synapsefs.skeleton:run
+#     fibonacci = synapsefs.skeleton:run
 # And any other entry points, for example:
 fs.opener =
-    syn = fs.opener.synapsefs:SynapseFSOpener
+    syn = synapsefs.opener:SynapseFSOpener
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
@@ -103,7 +103,7 @@ fs.opener =
 # CAUTION: --cov flags may prohibit setting breakpoints while debugging.
 #          Comment those flags to avoid this pytest issue.
 addopts =
-    --cov fs-synapsefs --cov-report term-missing --cov-report xml
+    --cov synapsefs --cov-report term-missing --cov-report xml
     --verbose
 norecursedirs =
     dist
