--- src/fs/opener/__init__.py
@@ -1,3 +1,0 @@
-from fs.opener.open_parent_fs import open_parent_fs
-
-__all__ = ["open_parent_fs"]
