+++ src/synapsefs/__init__.py
@@ -12,10 +12,11 @@ finally:
 
 import logging
 
-from fs.synapsefs.synapsefs import SynapseFS
+from synapsefs.open_parent_fs import open_parent_fs
+from synapsefs.synapsefs import SynapseFS
 
 # Set default logging handler to avoid "No handler found" warnings
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logging.captureWarnings(True)
 
-__all__ = ["SynapseFS"]
+__all__ = ["SynapseFS", "open_parent_fs"]
