+++ src/synapsefs/opener.py
@@ -2,7 +2,8 @@ import os
 
 from fs.opener import Opener
 from fs.opener.parse import ParseResult
-from fs.synapsefs.synapsefs import SynapseFS
+
+from synapsefs.synapsefs import SynapseFS
 
 
 class SynapseFSOpener(Opener):
