+++ tests/conftest.py
@@ -6,6 +6,8 @@ from uuid import uuid4
 
 import pytest
 
+from synapsefs import SynapseFS
+
 UUID = str(uuid4())
 USER = getuser()
 UTCTIME = datetime.now().isoformat(" ", "seconds").replace(":", ".")
@@ -14,3 +16,8 @@ RUNID = f"{USER} - {UTCTIME} - {UUID}"  # Valid characters: [A-Za-z0-9 .+'()_-]
 
 def pytest_configure():
     pytest.RUNID = RUNID  # type: ignore
+
+
+@pytest.fixture(scope="session")
+def synapse_fs():
+    yield SynapseFS()
